chore: rename train to trainer
