+++ src/features/trainer/components/TrainGerman/TrainGerman.tsx
