+++ src/features/trainer/components/TrainPhoneticPersian/TrainPhoneticPersian.tsx
