+++ src/features/trainer/model/learningProgress.ts
