+++ src/features/trainer/slice/configuration.ts
