+++ src/features/trainer/slice/selectRandom.spec.ts
