+++ src/features/trainer/slice/selectRandom.ts
